html

+++ src/main/java/com/example/fpoly_stadium/repository/DichVuSanBongRepository.java
@@ -3,5 +3,5 @@ package com.example.fpoly_stadium.repository;
 import com.example.fpoly_stadium.entity.san.DichVuSanBong;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DichVuSanBongRepository extends JpaRepository<DichVuSanBong, Long> {
+public interface DichVuSanBongRepository extends JpaRepository<DichVuSanBong, Integer> {
 }
