+++ src/main/java/com/example/fpoly_stadium/repository/DoUongRepository.java
@@ -5,5 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DoUongRepository extends JpaRepository<DoUong, Long> {
+public interface DoUongRepository extends JpaRepository<DoUong, Integer> {
+
 }
