+++ src/main/java/com/example/fpoly_stadium/repository/HinhThucThanhToanRepository.java
@@ -3,5 +3,5 @@ package com.example.fpoly_stadium.repository;
 import com.example.fpoly_stadium.entity.hoaDon.HinhThucThanhToan;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HinhThucThanhToanRepository extends JpaRepository<HinhThucThanhToan, Long> {
+public interface HinhThucThanhToanRepository extends JpaRepository<HinhThucThanhToan, Integer> {
 }
