+++ src/main/java/com/example/fpoly_stadium/repository/LoaiSanRepository.java
@@ -3,5 +3,5 @@ package com.example.fpoly_stadium.repository;
 import com.example.fpoly_stadium.entity.san.LoaiSan;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LoaiSanRepository extends JpaRepository<LoaiSan, Long> {
+public interface LoaiSanRepository extends JpaRepository<LoaiSan, Integer> {
 }
