+++ src/main/java/com/example/fpoly_stadium/repository/NgayTrongTuanRepository.java
@@ -3,5 +3,5 @@ package com.example.fpoly_stadium.repository;
 import com.example.fpoly_stadium.entity.san.NgayTrongTuan;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NgayTrongTuanRepository extends JpaRepository<NgayTrongTuan, Long> {
+public interface NgayTrongTuanRepository extends JpaRepository<NgayTrongTuan, Integer> {
 }
