+++ src/main/java/com/example/fpoly_stadium/repository/PhieuGiamGiaChiTietRepository.java
@@ -3,5 +3,5 @@ package com.example.fpoly_stadium.repository;
 import com.example.fpoly_stadium.entity.hoaDon.PhieuGiamGiaChiTiet;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PhieuGiamGiaChiTietRepository extends JpaRepository<PhieuGiamGiaChiTiet, Long> {
+public interface PhieuGiamGiaChiTietRepository extends JpaRepository<PhieuGiamGiaChiTiet, Integer> {
 }
