+++ src/main/java/com/example/fpoly_stadium/repository/SanBongRepository.java
@@ -3,5 +3,5 @@ package com.example.fpoly_stadium.repository;
 import com.example.fpoly_stadium.entity.san.SanBong;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SanBongRepository extends JpaRepository<SanBong, Long> {
+public interface SanBongRepository extends JpaRepository<SanBong, Integer> {
 }
