+++ src/main/java/com/example/fpoly_stadium/repository/SanCaRepository.java
@@ -3,5 +3,5 @@ package com.example.fpoly_stadium.repository;
 import com.example.fpoly_stadium.entity.san.SanCa;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SanCaRepository extends JpaRepository<SanCa, Long> {
+public interface SanCaRepository extends JpaRepository<SanCa, Integer> {
 }
